# perf: memoize DqElement