--- a/lib/core/utils/dq-element.js
+++ b/lib/core/utils/dq-element.js
@@ -4,6 +4,7 @@ import getXpath from './get-xpath';
 import getNodeFromTree from './get-node-from-tree';
 import AbstractVirtualNode from '../base/virtual-node/abstract-virtual-node';
 import cache from '../base/cache';
+import memoize from './memoize';
 
 const CACHE_KEY = 'DqElm.RunOptions';
 
@@ -36,7 +37,10 @@ function getSource(element) {
  * @param {Object} options Propagated from axe.run/etc
  * @param {Object} spec Properties to use in place of the element when instantiated on Elements from other frames
  */
-function DqElement(elm, options = null, spec = {}) {
+const DqElement = memoize(function DqElement(elm, options, spec) {
+  options ??= null;
+  spec ??= {};
+
   if (!options) {
     options = cache.get(CACHE_KEY) ?? {};
   }
@@ -82,7 +86,9 @@ function DqElement(elm, options = null, spec = {}) {
   if (!axe._audit.noHtml) {
     this.source = this.spec.source ?? getSource(this._element);
   }
-}
+
+  return this;
+});
 
 DqElement.prototype = {
   /**